Rewrite plot function
The exponential moving average was neigh impossible to initialize in a
reasonable way. Simply initializing it with the first value would, of
course, cause that value to be overrepresented. However, initializing it
with the mean of the first several values causes strictly increasing
functions to be plotted as if they descended then ascended.